requirements.txt
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
dbus-python==1.2.16
distro-info==1.0
dnspython==2.2.1
fastapi==0.75.0
gyp==0.1
h11==0.13.0
httptools==0.4.0
idna==3.3
pycparser==2.21
pycurl==7.43.0.6
pydantic==1.9.0
PyGObject==3.38.0
pymongo==4.0.2
python-apt==2.2.1
python-dotenv==0.19.2
PyYAML==6.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing-extensions==4.1.1
unattended-upgrades==0.1
urllib3==1.26.8
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8
websockets==10.2
